Clean-up Generic.cpp
Also applied some windows.h constants over raw values

+++ src/TextScr.h
@@ -55,7 +55,7 @@ struct TEXT_SCRIPT
 
 BOOL InitTextScript2(void);
 void EndTextScript(void);
-void EncryptionBinaryData2(unsigned char *pData, int size);
+void EncryptionBinaryData2(unsigned char *pData, long size);
 BOOL LoadTextScript2(const char *name);
 BOOL LoadTextScript_Stage(const char *name);
 void GetTextScriptPath(char *path);
